+++ src/components/Calender/CalenderInner.tsx
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
-import { formatMonthDate } from '../../utils/dateFormat';
-import CalenderData from '../CalenderData';
+import ColumnHeader from './ColumnHeader';
 import type Calender from '../../factory/Calender/Caldender';
 
 interface Props {
@@ -10,61 +9,44 @@ interface Props {
 
 const days = ['SUN', 'MON', 'TUE', 'WED', 'THU', 'FRI', 'SAT'];
 
-const CalenderInner = ({ calender, onClickItem }: Props) => {
-  const date = new Date(Date.now());
-
-  return (
-    <GridContainer>
-      <Row>
-        {days.map((day, index) => (
-          <Title key={day} isWeekend={index === 0 || index === 6}>
-            {day}
-          </Title>
-        ))}
-      </Row>
-      <Container>
-        {calender.calender.map(({ id, week }) => (
-          <Row key={id}>
-            {week.map((day, index) => (
-              <Column
-                key={day.id}
-                isWeekend={index === 0 || index === 6}
-                onClick={onClickItem}
-              >
-                <ColumnHeader>
-                  <CalenderDate
-                    isCurrentDate={date.getDate() === day.date}
-                    isCurrentMonth={date.getMonth() === day.month}
-                    isCurrentYear={date.getFullYear() === day.year}
-                    isUserSetMonth={calender.month === day.month}
-                  >
-                    {day.date === 1
-                      ? formatMonthDate(day.month, day.date)
-                      : day.date}
-                  </CalenderDate>
-                </ColumnHeader>
-              </Column>
-            ))}
-            {/** TODO: 여기에 데이터 들어가야 함, props으로 data 전달 받아서 표현할 것 */}
-            <CalenderData key={`data-${id}`} />
-          </Row>
-        ))}
-      </Container>
-    </GridContainer>
-  );
-};
+const CalenderInner = ({ calender, onClickItem }: Props) => (
+  <GridContainer>
+    <Row>
+      {days.map((day, index) => (
+        <Title key={day} isWeekend={index === 0 || index === 6}>
+          {day}
+        </Title>
+      ))}
+    </Row>
+    <Container>
+      {calender.calender.map(({ id, week }) => (
+        <Row key={id}>
+          {week.map((day, index) => (
+            <Column
+              key={day.id}
+              isWeekend={index === 0 || index === 6}
+              onClick={onClickItem}
+            >
+              <ColumnHeader
+                calenderMonth={calender.month}
+                date={day.date}
+                month={day.month}
+                year={day.year}
+              />
+            </Column>
+          ))}
+          {/** TODO: 여기에 데이터 들어가야 함, props으로 data 전달 받아서 표현할 것 */}
+          {/* <CalenderData key={`data-${id}`} /> */}
+        </Row>
+      ))}
+    </Container>
+  </GridContainer>
+);
 
 export default CalenderInner;
 
 interface CalenderColumnProps {
   isWeekend: boolean;
-}
-
-interface CalenderDateProps {
-  isUserSetMonth: boolean;
-  isCurrentMonth: boolean;
-  isCurrentYear: boolean;
-  isCurrentDate: boolean;
 }
 
 const Title = styled.div<CalenderColumnProps>`
@@ -108,42 +90,6 @@ const Column = styled.div<CalenderColumnProps>`
   &:hover {
     background-color: rgba(233, 233, 231, 0.3);
   }
-`;
-
-const ColumnHeader = styled.div`
-  display: flex;
-  // TODO: 나중에 지울 것
-  justify-content: flex-end;
-  padding-bottom: 0.5rem;
-`;
-
-const CalenderDate = styled.span<CalenderDateProps>`
-  display: inline-block;
-  min-width: 22px;
-  min-height: 22px;
-  border-radius: 50%;
-
-  color: ${({
-    isCurrentYear,
-    isCurrentMonth,
-    isCurrentDate,
-    isUserSetMonth,
-  }) => {
-    if (isCurrentYear && isCurrentMonth && isCurrentDate) {
-      return `#ffffff`;
-    }
-
-    if (!isUserSetMonth) {
-      return `rgba(55, 53, 47, 0.5)`;
-    }
-
-    return ``;
-  }};
-
-  background-color: ${({ isCurrentYear, isCurrentMonth, isCurrentDate }) =>
-    isCurrentYear && isCurrentMonth && isCurrentDate ? `rgb(235, 87, 87)` : ''};
-
-  font-size: 14px;
 `;
 
 const Container = styled.div`
